Refactor
-consolidated all 3 scripts into one python file for easier installation
-loading BRM_UVTools.py will now add buttons to the Shading/UVs section in the toolbar. From here you can run the tools or assign shortcuts
-Still a few usability issues to work out, best to assign hotkeys for now
